requirements.txt
aiogram==3.4.1
mysql-connector-python==8.3.0
Pillow==10.2.0
aiohttp==3.9.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.20.9  # Опционально, для ускорения fuzzywuzzy